\name{getStopwords}
\alias{getStopwords}

\title{Get list of common uninformative words}

\description{Get list of common uninformative words}

\usage{
    getStopwords(which = 'stopwords')
}

\arguments{
    \item{which}{A character among < stopwords >, < adjectives > or < adverbs >}
}

\value{A vector of english words.}

\seealso{
    \code{\link{removeWords}}
}

\author{Nicolas Casajus}

\examples{
### See https://github.com/ahasverus/rscimap
}
